# swagger-api-spec.yaml
openapi: 3.0.3
info:
  title: Agent Monitor Data Source API
  description: API for implementing the IDataSource interface with complete CRUD operations for all entity types.
  version: 1.0.0
servers:
  - url: https://api.agent-monitor.example.com/v1
    description: Main production server
  - url: https://api.agent-monitor-staging.example.com/v1
    description: Staging server
...
paths:
...
  /messages:
    get:
      summary: Get messages
      description: Retrieve messages by IDs or all messages if no IDs provided
      operationId: getMessages
      tags:
        - Messages
      parameters:
        - in: query
          name: ids
          description: Comma-separated list of message IDs
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a message
      description: Create a new message
      operationId: createMessage
      tags:
        - Messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageInput'
      responses:
        '201':
          description: Successfully created message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /messages/{id}:
    get:
      summary: Get message by ID
      description: Retrieve a specific message by its ID
      operationId: getMessageById
      tags:
        - Messages
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update message
      description: Update an existing message
      operationId: updateMessage
      tags:
        - Messages
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageUpdate'
      responses:
        '200':
          description: Successfully updated message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete message
      description: Delete an existing message
      operationId: deleteMessage
      tags:
        - Messages
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Successfully deleted message
          content:
            application/json:
              schema:
                type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /aiagents/{aiAgentId}/conversations:
    get:
      summary: Get conversations by AI agent ID
      description: Retrieve all conversations for a specific AI agent
      operationId: getConversationsByAIAgentId
      tags:
        - Conversations
      parameters:
        - name: aiAgentId
          in: path
          description: ID of the AI agent
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /aiagents:
    get:
      summary: Get AI agents
      description: Retrieve AI agents by IDs or all AI agents if no IDs provided
      operationId: getAIAgents
      tags:
        - AIAgents
      parameters:
        - in: query
          name: ids
          description: Comma-separated list of AI agent IDs
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/AIAgent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create an AI agent
      description: Create a new AI agent
      operationId: createAIAgent
      tags:
        - AIAgents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIAgentInput'
      responses:
        '201':
          description: Successfully created AI agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIAgent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /aiagents/{id}:
    get:
      summary: Get AI agent by ID
      description: Retrieve a specific AI agent by its ID
      operationId: getAIAgentById
      tags:
        - AIAgents
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIAgent'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update AI agent
      description: Update an existing AI agent
      operationId: updateAIAgent
      tags:
        - AIAgents
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIAgentUpdate'
      responses:
        '200':
          description: Successfully updated AI agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIAgent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete AI agent
      description: Delete an existing AI agent
      operationId: deleteAIAgent
      tags:
        - AIAgents
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Successfully deleted AI agent
          content:
            application/json:
              schema:
                type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /aiagents/status/{status}:
    get:
      summary: Get AI agents by status
      description: Retrieve all AI agents with a specific status
      operationId: getAIAgentsByStatus
      tags:
        - AIAgents
      parameters:
        - name: status
          in: path
          description: Status of the AI agent (active, inactive, or training)
          required: true
          schema:
            type: string
            enum: ['active', 'inactive', 'training']
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AIAgent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # User endpoints
...
  /system/initialize:
    post:
      summary: Initialize data source
      description: Initialize the data source and prepare it for use
      operationId: initialize
      tags:
        - System
      responses:
        '200':
          description: Successfully initialized data source
        '500':
          $ref: '#/components/responses/InternalServerError'
  /system/save:
    post:
      summary: Save data
      description: Save all data to persistent storage
      operationId: saveData
      tags:
        - System
      responses:
        '200':
          description: Successfully saved data
        '500':
          $ref: '#/components/responses/InternalServerError'
  /system/cache/clear:
    post:
      summary: Clear cache
      description: Clear all cached data
      operationId: clearCache
      tags:
        - System
      responses:
        '200':
          description: Successfully cleared cache
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    IdParam:
      name: id
      in: path
      description: ID of the resource
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the message
        content:
          type: string
          description: Content of the message
        timestamp:
          type: string
          format: date-time
          description: When the message was sent
        sender:
          type: string
          enum: ['user', 'ai']
          description: Who sent the message
        senderName:
          type: string
          description: Name of the sender
        messageType:
          type: string
          enum: ['text', 'attachment', 'system']
          description: Type of message
        readStatus:
          type: boolean
          description: Whether the message has been read
        metadata:
          type: object
          properties:
            tags:
              type: array
              items:
                type: string
              description: Tags associated with the message
            priority:
              type: string
              enum: ['low', 'medium', 'high']
              description: Priority of the message
            confidence:
              type: string
              description: Confidence level of the message (for AI responses)
      required: [id, content, timestamp, sender, senderName, messageType, readStatus, metadata]
    Conversation:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the conversation
        userId:
          type: string
          description: ID of the user participating in the conversation
        userName:
          type: string
          description: Name of the user participating in the conversation
        aiAgentId:
          type: string
          description: ID of the AI agent participating in the conversation
        aiAgentName:
          type: string
          description: Name of the AI agent participating in the conversation
        aiAgentType:
          type: string
          description: Type of the AI agent participating in the conversation
        status:
          type: string
          enum: ['active', 'closed']
          description: Status of the conversation
        conclusion:
          type: string
          enum: ['successful', 'unsuccessful', 'pending']
          description: Conclusion of the conversation
        startTimestamp:
          type: string
          format: date-time
          description: When the conversation started
        endTimestamp:
          type: string
          format: date-time
          description: When the conversation ended (if closed)
        messages:
          type: array
          items:
            type: string
          description: IDs of messages in the conversation
        tags:
          type: array
          items:
            type: string
          description: Tags associated with the conversation
        resolutionNotes:
          type: string
          description: Notes about how the conversation was resolved
        priority:
          type: string
          enum: ['low', 'medium', 'high']
          description: Priority of the conversation
        duration:
          type: string
          description: Duration of the conversation
        messageCount:
          type: integer
          description: Number of messages in the conversation
        confidence:
          type: string
          description: Overall confidence of the AI agent in the conversation
        conversationTimestamp:
          type: string
          format: date-time
          description: Timestamp of the first message
      required: [id, userId, userName, aiAgentId, aiAgentName, aiAgentType, status, conclusion, startTimestamp, messages, tags, priority, duration, messageCount, confidence]
...
    AIAgent:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the AI agent
        name:
          type: string
          description: Name of the AI agent
        model:
          type: string
          description: Model used by the AI agent
        status:
          type: string
          enum: ['active', 'inactive', 'training']
          description: Status of the AI agent
        conversationsProcessed:
          type: integer
          description: Number of conversations processed by the AI agent
        successRate:
          type: string
          description: Success rate of the AI agent
        avgResponseTime:
          type: string
          description: Average response time of the AI agent
        lastActive:
          type: string
          format: date-time
          description: When the AI agent was last active
        capabilities:
          type: array
          items:
            type: string
          description: Capabilities of the AI agent
        specializations:
          type: array
          items:
            type: string
          description: Specializations of the AI agent
      required: [id, name, model, status, conversationsProcessed, successRate, avgResponseTime, lastActive]
...
    MessageInput:
      type: object
      properties:
        content:
          type: string
          description: Content of the message
        timestamp:
          type: string
          format: date-time
          description: When the message was sent
        sender:
          type: string
          enum: ['user', 'ai']
          description: Who sent the message
        senderName:
          type: string
          description: Name of the sender
        messageType:
          type: string
          enum: ['text', 'attachment', 'system']
          description: Type of message
        readStatus:
          type: boolean
          description: Whether the message has been read
        metadata:
          type: object
          properties:
            tags:
              type: array
              items:
                type: string
              description: Tags associated with the message
            priority:
              type: string
              enum: ['low', 'medium', 'high']
              description: Priority of the message
            confidence:
              type: string
              description: Confidence level of the message (for AI responses)
      required: [content, timestamp, sender, senderName, messageType, readStatus, metadata]
...
    AIAgentInput:
      type: object
      properties:
        name:
          type: string
          description: Name of the AI agent
        model:
          type: string
          description: Model used by the AI agent
        status:
          type: string
          enum: ['active', 'inactive', 'training']
          description: Status of the AI agent
        conversationsProcessed:
          type: integer
          description: Number of conversations processed by the AI agent
        successRate:
          type: string
          description: Success rate of the AI agent
        avgResponseTime:
          type: string
          description: Average response time of the AI agent
        lastActive:
          type: string
          format: date-time
          description: When the AI agent was last active
        capabilities:
          type: array
          items:
            type: string
          description: Capabilities of the AI agent
        specializations:
          type: array
          items:
            type: string
          description: Specializations of the AI agent
      required: [name, model, status, conversationsProcessed, successRate, avgResponseTime, lastActive]
...
    MessageUpdate:
      type: object
      properties:
        content:
          type: string
          description: Content of the message
        readStatus:
          type: boolean
          description: Whether the message has been read
        metadata:
          type: object
          properties:
            tags:
              type: array
              items:
                type: string
              description: Tags associated with the message
            priority:
              type: string
              enum: ['low', 'medium', 'high']
              description: Priority of the message
            confidence:
              type: string
              description: Confidence level of the message (for AI responses)
...
    AIAgentUpdate:
      type: object
      properties:
        name:
          type: string
          description: Name of the AI agent
        model:
          type: string
          description: Model used by the AI agent
        status:
          type: string
          enum: ['active', 'inactive', 'training']
          description: Status of the AI agent
        conversationsProcessed:
          type: integer
          description: Number of conversations processed by the AI agent
        successRate:
          type: string
          description: Success rate of the AI agent
        avgResponseTime:
          type: string
          description: Average response time of the AI agent
        lastActive:
          type: string
          format: date-time
          description: When the AI agent was last active
        capabilities:
          type: array
          items:
            type: string
          description: Capabilities of the AI agent
        specializations:
          type: array
          items:
            type: string
          description: Specializations of the AI agent
...
    Error:
      type: object
      properties:
        code:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message
        details:
          type: object
          additionalProperties: true
          description: Additional details about the error
      required: [code, message]
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []